api/requirements.txt
flask==2.3.3
Flask-CORS==3.0.10
sigma-cli==0.7.6
pysigma==0.9.11
pySigma-backend-azure @ git+https://github.com/sifex/pySigma-backend-azure.git@9c35b78cd4d4891e1b46ccf21ed555ffb10b96e6
pysigma-backend-carbonblack==0.1.4
pysigma-backend-cortexxdr==0.1.1
pysigma-backend-elasticsearch==1.0.5
pySigma-backend-loki @ git+https://github.com/grafana/pySigma-backend-loki.git@452aa0d8bb096bbabdca5a83a884c45662dec666
pySigma-backend-microsoft365defender @ git+https://github.com/AttackIQ/pySigma-backend-microsoft365defender.git@731db4b6c7cbab8898973b4350d93268e135c495
pysigma-backend-opensearch==1.0.0
pysigma-backend-qradar==0.3.3
pySigma-backend-QRadar-AQL==0.2.3
pysigma-backend-sentinelone==0.1.2
pysigma-backend-sentinelone-pq==0.1.1
pysigma-backend-splunk==1.0.2
pySigma-backend-stix @ git+https://github.com/barvhaim/pySigma-backend-stix@0d7c05e187249c26f5abb99f63bf839c78705d1e
pysigma-pipeline-crowdstrike==1.0.0
pysigma-pipeline-sysmon==1.0.2
pysigma-pipeline-windows==1.1.0